requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.32
Flask-Mail==0.9.1
celery==5.3.6
redis==5.0.3
python-dotenv==1.0.1
gunicorn==22.0.0
